Fix bug where M-x live-preview could cause error
M-x live-preview uses the old value of live-preview-command as the
default input string in the minibuffer prompt. When the value of
live-preview-command is a function (not a string), Emacs won't be able
to insert it into the minibuffer and an error occurs.

Fix by ignoring the existing value if it's a function.