Try to fix occasional garbled output
There was a bug where rendering the preview would occassionally clear
the source buffer of prepend some output from the preview
command (nroff) to the source buffer.

Try to prevent that from happening by ensuring as late as possible
that `current-buffer` is the preview buffer before clearing or
inserting any output.